backend/requirements.txt
flask
websocket-client

flask-socketio
flask-cors

